docs/modulos_preprocesamiento_espectral/1_correccion_geometrica/requirements.txt
numpy>=1.22.4
pandas>=2.0.3
opencv-python>=4.8.1.78
rasterio>=1.3.7
tifffile>=2023.4.12
Pillow>=10.0.0
scikit-image>=0.21.0